requirements.txt
colorlog>=6.10.1,<6.11.0
homeassistant>=2025.6.0,<2025.7.0
pytest-homeassistant-custom-component>=0.13.0,<0.14.0
pip>=21.3.1
pre-commit>=4.3.0,<4.4.0
ruff>=0.14.1,<0.15.0
